requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
httpx==0.27.0
websockets==12.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
